Rename components now that we've lost legacy ones.
