+++ packages/components/src/index.ts
@@ -41,6 +41,6 @@ export { default as DataTable } from './local/organisms/data-table'
 export { default as GameSetup } from './local/templates/game-setup'
 export { default as ChatMessageSetup } from './local/templates/chat-message-setup'
 export { default as ChatEntryForm } from './local/form-elements/chat-entry-form'
-export { default as CoaStatusBoard2 } from './local/coa-status-board-2'
+export { default as CoaStatusBoard } from './local/coa-status-board'
 export { default as HeartbeatChecker } from './local/atoms/heartbeat-checker'
 export { default as Confirm } from './local/atoms/confirm'
