+++ packages/components/src/local/coa-status-board/index.spec.tsx
@@ -5,7 +5,7 @@ import renderer from 'react-test-renderer'
 import { ForceRole, MessageCustom, Role, ChannelCollab } from '@serge/custom-types'
 import { CUSTOM_MESSAGE, CollaborativeMessageStates, CHANNEL_COLLAB, InitialStates, CollaborativePermission, CollaborativeMessageStates2, PARTICIPANT_COLLAB } from '@serge/config'
 import { forces, MessageTemplatesMockByKey, WargameMock } from '@serge/mocks'
-import CoaStatusBoard2 from './index'
+import CoaStatusBoard from './index'
 import moment from 'moment'
 
 /* Import mock data */
@@ -212,7 +212,7 @@ it('CoaStatusBoard renders correctly', () => {
 
   const tree = renderer
     .create(
-      <CoaStatusBoard2
+      <CoaStatusBoard
         currentWargame='wargame-test'
         templates={MessageTemplatesMockByKey}
         forces={forces}
