+++ packages/components/src/local/coa-status-board/types/props.d.ts
