+++ packages/components/src/local/molecules/channel-coa-message-detail/README.md
