+++ packages/components/src/local/molecules/channel-coa-message-detail/styles.module.scss
