+++ packages/components/src/local/molecules/channel-coa-message-detail/types/props.d.ts
